Update pythoncalculator.py
Changed int to float

+++ pythoncalculator.py
@@ -3,8 +3,9 @@
 print("Would you like to do some addition?")
 answer = input()
 if answer == "yes":
-    num1 = int(input("Please enter the first number you would like to add together: "))
-    num2 = int(input("Please enter the second number you would like to add together: "))
+    # Changed int to float here
+    num1 = float(input("Please enter the first number you would like to add together: "))
+    num2 = float(input("Please enter the second number you would like to add together: "))
     print(num1 + num2)
 else:
     print("Okay, understood")
@@ -13,8 +14,9 @@ else:
 print("Would you like to do some multiplication?")
 answer = input()
 if answer == "yes":
-    num1 = int(input("Please enter first number you would like to times together: "))
-    num2 = int(input("Please enter second number you would like to times together: "))
+    # Changed int to float here
+    num1 = float(input("Please enter first number you would like to times together: "))
+    num2 = float(input("Please enter second number you would like to times together: "))
     print(num1 * num2)
 else:
     print("Okay, understood")
@@ -23,8 +25,9 @@ else:
 print("Would you like to do some division?")
 answer = input()
 if answer == "yes":
-    num1 = int(input("Please enter the first number you would like to divide: "))
-    num2 = int(input("Please enter the second number you would like to divide from the first number: "))
+    # Changed int to float here
+    num1 = float(input("Please enter the first number you would like to divide: "))
+    num2 = float(input("Please enter the second number you would like to divide from the first number: "))
     print(num1 / num2)
 else:
     print("Okay understood")
@@ -33,8 +36,9 @@ else:
 print("Would you like to do some subtraction?")
 answer = input()
 if answer == "yes":
-    num1 = int(input("Please enter the first number you would like to subtract: "))
-    num2 = int(input("Please enter the second number you would like to subtract from the first: "))
+    # Changed int to float here
+    num1 = float(input("Please enter the first number you would like to subtract: "))
+    num2 = float(input("Please enter the second number you would like to subtract from the first: "))
     print(num1 - num2)
 else:
     print("Okay, understood")
